test: remove v1 tests
