+++ package.json
@@ -12,7 +12,7 @@
     "test:lint": "eslint .",
     "test:unit": "nyc --report-dir=test/coverage --reporter=lcov mocha \"test/unit/**/*.spec.js\" --require test/test-helpers.js -- $npm_package_config_reporter/unit.json",
     "test:ui": "npm run test:ui:firefox && npm run test:ui:webkit && npm run test:ui:chromium",
-    "test:ui:browser": "mocha \"test/ui/v2/**/*.spec.js\" --require test/test-helpers.js test/ui/ui-test-helpers.js --timeout 10000",
+    "test:ui:browser": "mocha \"test/ui/**/*.spec.js\" --require test/test-helpers.js test/ui/ui-test-helpers.js --timeout 10000",
     "test:ui:chromium": "npm run test:ui:browser -- --browser-engine=chromium $npm_package_config_reporter/chromium.json",
     "test:ui:firefox": "npm run test:ui:browser -- --browser-engine=firefox $npm_package_config_reporter/firefox.json",
     "test:ui:webkit": "npm run test:ui:browser -- --browser-engine=webkit $npm_package_config_reporter/webkit.json",
