+++ test/ui/04 - applicant-details/content.spec.js
