+++ test/ui/07 - query/query.spec.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const config = require('../../ui-test-config');
+const config = require('../ui-test-config');
 
 describe('/query', () => {
