+++ test/ui/11 - errors/page-not-found.spec.js
@@ -1,0 +1,1 @@
+'use strict';
